# replace czlib with go-libdeflate