--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -10,28 +10,27 @@ jobs:
   build-and-test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       
       - name: Setup Go
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v4
         with:
           go-version: '1.19.13'
+          cache: false
       
       - name: Install dependencies
         run: |
           go version
-          go get -u golang.org/x/lint/golint
 
       - name: Run build
         run: go build . 
       
+      - name: Run tests
+        run: go test -v -coverprofile=profile.cov ./...
+
       - name: Run vet & lint
         run: |
           go vet .
-          golint .
-    
-      - name: Run tests
-        run: go test -v -coverprofile=profile.cov ./...
    
       - name: codecov
         uses: codecov/codecov-action@v1