--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/paulmach/osm
 go 1.13
 
 require (
-	github.com/datadog/czlib v0.0.0-20160811164712-4bc9a24e37f2
+	github.com/4kills/go-libdeflate/v2 v2.0.3
 	github.com/paulmach/orb v0.1.3
 	github.com/paulmach/protoscan v0.2.1
 	golang.org/x/time v0.0.0-20190921001708-c4c64cad1fd0