--- a/go.mod
+++ b/go.mod
@@ -1,12 +1,13 @@
 module github.com/paulmach/osm
 
-go 1.13
+go 1.19
 
 require (
 	github.com/4kills/go-libdeflate/v2 v2.0.3
-	github.com/paulmach/orb v0.1.3
+	github.com/paulmach/orb v0.10.0
 	github.com/paulmach/protoscan v0.2.1
 	golang.org/x/time v0.0.0-20190921001708-c4c64cad1fd0
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 	google.golang.org/protobuf v1.27.1
 )
+
+require go.mongodb.org/mongo-driver v1.12.1 // indirect