--- a/osmpbf/zlib_go.go
+++ b/osmpbf/zlib_go.go
@@ -1,3 +1,4 @@
+//go:build !cgo
 // +build !cgo
 
 package osmpbf
@@ -5,9 +6,29 @@ package osmpbf
 import (
 	"bytes"
 	"compress/zlib"
-	"io"
+	"fmt"
 )
 
-func zlibReader(data []byte) (io.ReadCloser, error) {
-	return zlib.NewReader(bytes.NewReader(data))
+func decompress(in []byte, size int, data []byte) ([]byte, error) {
+	r, err := zlib.NewReader(bytes.NewReader(in))
+	if err != nil {
+		return nil, err
+	}
+
+	// using the bytes.Buffer allows for the preallocation of the necessary space.
+	l := size + bytes.MinRead
+	if cap(data) < int(l) {
+		data = make([]byte, 0, l+l/10)
+	} else {
+		data = data[:0]
+	}
+	buf := bytes.NewBuffer(data)
+	if _, err = buf.ReadFrom(r); err != nil {
+		return nil, err
+	}
+
+	if buf.Len() != int(size) {
+		return nil, fmt.Errorf("raw blob data size %d but expected %d", buf.Len(), size)
+	}
+	return buf.Bytes(), nil
 }