Performance can be improved by using libdeflate, which is better suited 
zlib uncompression. Care was taken to reuse the buffer "data".

Since Workflow had errors I renovated it a bit. However I did not use
https://golangci-lint.run since it showed too many nitpicks.

Benchmark improvements from czlib to libdeflate:
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkLondon-8                      296011237     275254714     -7.01%
BenchmarkLondon_withFiltersTrue-8      282195335     263935960     -6.47%
BenchmarkLondon_withFiltersFalse-8     231823251     200477972     -13.52%
BenchmarkLondon_nodes-8                216665255     180614979     -16.64%
BenchmarkLondon_ways-8                 167160662     140700970     -15.83%
BenchmarkLondon_relations-8            99489502      75263200      -24.35%

benchmark                              old allocs     new allocs     delta
BenchmarkLondon-8                      4811462        4808526        -0.06%
BenchmarkLondon_withFiltersTrue-8      4811463        4808515        -0.06%
BenchmarkLondon_withFiltersFalse-8     1367676        1364724        -0.22%
BenchmarkLondon_nodes-8                3398498        3395559        -0.09%
BenchmarkLondon_ways-8                 1799045        1796099        -0.16%
BenchmarkLondon_relations-8            463103         460152         -0.64%

benchmark                              old bytes     new bytes     delta
BenchmarkLondon-8                      938187018     924789892     -1.43%
BenchmarkLondon_withFiltersTrue-8      938186796     924787588     -1.43%
BenchmarkLondon_withFiltersFalse-8     379851201     366452840     -3.53%
BenchmarkLondon_nodes-8                632788308     619391213     -2.12%
BenchmarkLondon_ways-8                 451758725     438360054     -2.97%
BenchmarkLondon_relations-8            198025084     184626277     -6.77%```