@paulmach 2023-12-28
> Why not [`compress/zlib`](https://pkg.go.dev/compress/zlib)?

Speed. See the comment in the [osmpdb/README.md](https://github.com/paulmach/osm/tree/master/osmpbf#using-cgoczlib-for-decompression). 25% faster with the current lib. This change looks to make it even faster.