@paulmach 2023-12-28
I merged these change into https://github.com/paulmach/osm/pull/55 so that I could fix the issues.  I was unable to run locally due to some linker issues, will need to debug.